vehicles/serializers.py
from rest_framework import serializers

from vehicles.models import Manufacturer


class ManufacturerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Manufacturer
        fields = ['id', 'name']
